fix: modified popup height and included library
The popup looked too small and text felt too cramped, so I decided to raise the window's height. Then, the GetModuleFileNameA function was included in the windows.h header, which defined a macro that used a name already defined in the PDCurses library, so I just included the libloaderapi.h header on which it was declared.